requirements.txt
alembic==1.0.11
aniso8601==7.0.0
attrs==19.1.0
bcrypt==3.1.7
boto3==1.9.208
botocore==1.12.208
certifi==2019.6.16
cffi==1.12.3
Click==7.0
croniter==0.3.30
docutils==0.14
et-xmlfile==1.0.1
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-GoogleMaps==0.2.6
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-RQ2==18.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
Mako==1.0.13
MarkupSafe==1.1.1
numpy==1.17.0
openpyxl==2.6.2
pandas==0.25.0
pycparser==2.19
pyrsistent==0.15.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
redis==3.3.2
rq==1.1.0
rq-scheduler==0.9
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
uuid==1.30
Werkzeug==0.15.4
wincertstore==0.2
WTForms==2.2.1
xlrd==1.2.0
